backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
pypdf
sentence-transformers
tiktoken
docarray
langchain-text-splitters
chromadb
langchain-google-genai
python-dotenv
